refactor: centralize auth context
