+++ src/Pages/Dashboard.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import { useAuth } from '@/AuthContext';
+import { useAuth } from '@/context/AuthContext';
 import { fetchFromAPI } from '../lib/api';
 import type { Player } from '../types';
